- Updated FailureMechanismContributionNormChangeHandler.ChangeNorm's signature to pass along the norm, instead of the return period - Had to update all tests that create an AssessmentSection that was defined as a return period Related to Issue [WTI-1019]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@6953 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 00e291aa4d0d4d90e60e41432032bc2bf6e3d847
Former-commit-id: 85d3c72325fb1550f01ddf9ca8716034496aa175

+++ Ringtoets/Integration/src/Ringtoets.Integration.Forms/Views/IFailureMechanismContributionNormChangeHandler.cs
@@ -52,6 +52,6 @@ namespace Ringtoets.Integration.Forms.Views
         /// is <c>null</c>.</exception>
         /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="newNormValue"/>
         /// is an invalid norm value.</exception>
-        IEnumerable<IObservable> ChangeNorm(IAssessmentSection assessmentSection, int newNormValue);
+        IEnumerable<IObservable> ChangeNorm(IAssessmentSection assessmentSection, double newNormValue);
     }
 }
